feat(Action): convert axis data to float angle with deadzone limits
The AxesToAngleAction takes a horizontal and vertical axis and converts
it to an angle which can be used to determine the direction in which
a 2d axis is being pushed. The angle output can then be fed into
something to provide an action such as rotating the object to face
in a certain direction.