feat(AngleRangeToBoolean): add change angle unit and higher angle range
The angle unit can now be changed via the inspector and the angle range
can extend between -360f and 360f to ensure all angle unit types can
be catered for.

This now makes it possible to get a downward angle, which would be in
degrees (or radians) at 170 degrees <--> 200 degress, whereas this
isn't possible with signed units as -180 degress is the far left and
180 is the far right, but there is no simple way of knowing the down
without using the unsigned unit type.