feat(Prefabs): expose time multiplier on AxesToVector3Action
The AxesToVector3Action had a nested internal TimeComponentExtractor
that was used to extract a Time value to multiply the axis data by.

This has now been exposed on the Facade to make it easier to change
and the HowToGuide has been updated to show this needs setting to
Delta Time when doing axis movement.