feat(AxesToVector3Action): provide combined deadzone option
The deadzones were previously calculated per axis meaning the axis
data would not be reported for each individual axis. However, this
meant that circular motions outside of the initial axis would not
be correctly reported.

The solution is to provide a combined axis deadzone option that
implements a zonal axis deadzone that takes all axis deadzones into
consideration.