fix(Prefabs): ensure deadzone types are disabled by default
The Single DeadZone calculation and the Combined DeadZone calculation
both need to start disabled otherwise their internal code can run on
Enable due to a change in Zinnia how the `isActiveAndEnabled` check is
now done.

Previously, it was using `behaviour.isActiveAndEnabled` which will not
consider a script enabled until the `Start` method has been called so
anything in `OnEnable` wouldn't trigger. But this logic is misleading
so it was replaced with `gameObject.activeInHierarcy && this.enabled`
which will report a component being enabled in the `OnEnable`.

This was then highlighting the error in this prefab as it was running
code in the first frame because both components are enabled by default
and then it was disabling the relevant component. But by that time
the error had already occurred.

The fix is to ensure both calculation logic blocks are disabled by
default and then the code will enabled the correct one and run the
correct script without crashing into each other.